requirements.txt
Flask==2.3.3
selenium==4.15.2
Pillow==10.0.1
